module/power_domain: Extract state utility functions
The state related utility functions are now in their own file.
This helps to remove some of the clutter from the main module file.
