+++ module/power_domain/CMakeLists.txt
@@ -1,6 +1,6 @@
 #
 # Arm SCP/MCP Software
-# Copyright (c) 2021, Arm Limited and Contributors. All rights reserved.
+# Copyright (c) 2021-2023, Arm Limited and Contributors. All rights reserved.
 #
 # SPDX-License-Identifier: BSD-3-Clause
 #
@@ -13,4 +13,5 @@ target_include_directories(${SCP_MODULE_TARGET}
 target_sources(
     ${SCP_MODULE_TARGET}
     PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/src/mod_power_domain.c"
-            "${CMAKE_CURRENT_SOURCE_DIR}/src/power_domain_utils.c")
+            "${CMAKE_CURRENT_SOURCE_DIR}/src/power_domain_utils.c"
+            "${CMAKE_CURRENT_SOURCE_DIR}/src/power_domain_state_checks.c")
